+++ contents/BenefitTasks.js
@@ -1,0 +1,186 @@
+export const BenefitTasks = {
+  // general tasks
+  UpdateAccountInfoTask: {
+    task: 'updateAccountInfoTask',
+    icon: '/images/dashboard/account-info-icon.svg',
+    link: 'updateAccountInfoTaskLink',
+  },
+  TaxSlipTask: {
+    task: 'taxSlipTask',
+    icon: '/images/dashboard/oas-tax-slip-icon.svg',
+    link: 'taxSlipTaskLink',
+  },
+  TaxSlipMailingTask: {
+    task: 'taxSlipMailingTask',
+    icon: '/images/dashboard/oas-tax-slip-mailing-icon.svg',
+    link: 'taxSlipMailingLink',
+  },
+  // oas cpp tasks
+  AllPaymentsTask: {
+    task: 'allPaymentsTask',
+    icon: '/images/dashboard/oas-payment-icon.svg',
+    link: 'allPaymentsTaskLink',
+  },
+  StatusUpdateTask: {
+    task: 'statusUpdateTask',
+    icon: '/images/dashboard/oas-updates-message-icon.svg',
+    link: 'statusUpdateTaskLink',
+  },
+  RetirementIncomeTask: {
+    task: 'retirementIncomeTask',
+    icon: '/images/dashboard/oas-retirement-income-icon.svg',
+    link: 'retirementIncomeTaskLink',
+  },
+  ReconsiderationTask: {
+    task: 'reconsiderationTask',
+    icon: '/images/dashboard/oas-request-for-reconsideration-icon.svg ',
+    link: 'reconsiderationLink',
+  },
+  DelayOasPensionTask: {
+    task: 'delayOasPensionTask',
+    icon: '/images/dashboard/oas-delay-receiving-pension-icon.svg',
+    link: 'delayOasPensionTaskLink',
+  },
+  CppContributionTask: {
+    task: 'cppContributionTask',
+    icon: '/images/dashboard/oas-cpp-contributions-icon.svg',
+    link: 'cppContributionTaskLink',
+  },
+  TaxDeductionsTask: {
+    task: 'taxDeductionsTask',
+    icon: '/images/dashboard/oas-federal-tax-deductions-icon.svg',
+    link: 'taxDeductionsTaskLink',
+  },
+  GiveConsentTask: {
+    task: 'giveConsentTask',
+    icon: '/images/dashboard/oas-consent-icon.svg',
+    link: 'giveConsentTaskLink',
+  },
+  TaxSlipT4aAndNr4Task: {
+    task: 'taxSlipT4aAndNr4Task',
+    icon: '/images/dashboard/oas-tax-slip-icon.svg',
+    link: 'taxSlipTaskLink',
+  },
+  UploadMyDocuments: {
+    task: 'uploadMyDocuments',
+    icon: '/images/dashboard/ei-submit-documents-icon.svg',
+    link: 'uploadMyDocumentsLink',
+  },
+  ApplyForCppDisabilityBenefits: {
+    task: 'applyForCppDisabilityBenefits',
+    icon: '/images/dashboard/apply-for-benefit-icon-white.svg',
+    link: '/dashboard',
+  },
+  ApplyForCppRetirementPensions: {
+    task: 'applyForCppRetirementPensions',
+    icon: '/images/dashboard/apply-for-benefit-icon-white.svg',
+    link: '/dashboard',
+  },
+  EstimateMyMonthlyCppBenefits: {
+    task: 'estimateMyMonthlyCppBenefits',
+    icon: '/images/dashboard/oas-retirement-income-icon.svg',
+    link: '/dashboard',
+  },
+  ViewMyCppContributions: {
+    task: 'viewMyCppContributions',
+    icon: '/images/dashboard/oas-cpp-contributions-icon.svg',
+    link: '/dashboard',
+  },
+  ApplyForCppDeathBenefits: {
+    task: 'applyForCppDeathBenefits',
+    icon: '/images/dashboard/apply-for-benefit-icon-white.svg',
+    link: '/dashboard',
+  },
+  ApplyForGuranteedIncomeSupplement: {
+    task: 'applyForGuranteedIncomeSupplement',
+    icon: '/images/dashboard/apply-for-benefit-icon-white.svg',
+    link: '/dashboard',
+  },
+  ApplyForCppSurvivorsPensionAndChildrensBenefit: {
+    task: 'applyForCppSurvivorsPensionAndChildrensBenefit',
+    icon: '/images/dashboard/apply-for-benefit-icon-white.svg',
+    link: '/dashboard',
+  },
+  ApplyForOldAgeSecurity: {
+    task: 'applyForOldAgeSecurity',
+    icon: '/images/dashboard/apply-for-benefit-icon-white.svg',
+    link: '/dashboard',
+  },
+  // ei tasks
+  EiStatusUpdateTask: {
+    task: 'statusUpdateTask',
+    icon: '/images/dashboard/oas-updates-message-icon.svg',
+    taskLink: 'eiStatusAndMessages',
+  },
+  CompleteInsuranceReportTask: {
+    task: 'completeInsuranceReportTask',
+    icon: '/images/dashboard/ei-insurance-report-icon.svg',
+    taskLink: 'completeInsuranceReportLink',
+  },
+  ViewPaymentInfo: {
+    task: 'viewPaymentInfo',
+    icon: '/images/dashboard/ei-view-payment-info-icon.svg',
+    taskLink: 'viewPaymentInfoLink',
+  },
+  AccessCode: {
+    task: 'accessCode',
+    icon: '/images/dashboard/ei-access-code.svg',
+    taskLink: '/dashboard',
+  },
+  SubmitDocuments: {
+    task: 'submitDocuments',
+    icon: '/images/dashboard/ei-submit-documents-icon.svg',
+    taskLink: 'eiUploadDocuments',
+  },
+  ViewDocuments: {
+    task: 'viewDocuments',
+    icon: '/images/dashboard/ei-view-documents-icon.svg',
+    taskLink: 'viewDocumentsLink',
+  },
+  ViewLatestClaimTask: {
+    task: 'viewLatestClaimTask',
+    icon: '/images/dashboard/ei-latest-claim-icon.svg',
+    taskLink: 'viewLatestClaimLink',
+  },
+  ViewPastClaimsTask: {
+    task: 'viewPastClaimsTask',
+    icon: '/images/dashboard/ei-past-claims-icon.svg',
+    taskLink: 'viewPastClaimsLink',
+  },
+  SubmitEformsTask: {
+    task: 'submitEformsTask',
+    icon: '/images/dashboard/ei-submit-eform-icon.svg',
+    taskLink: 'submitFormsLink',
+  },
+  RegisterForAlerts: {
+    task: 'registerForAlerts',
+    icon: '/images/dashboard/ei-register-for-alert-icon.svg ',
+    taskLink: 'eiRegisterForAlerts',
+  },
+  RecordOfEmployment: {
+    task: 'recordOfEmployment',
+    icon: '/images/dashboard/ei-record-of-employment-icon.svg',
+    taskLink: 'recordOfEmployment',
+  },
+  ReportMistake: {
+    task: 'reportMistake',
+    icon: '/images/dashboard/ei-report-mistake-icon.svg ',
+    taskLink: 'eiReportMistake',
+  },
+  TaxSlipT4eTask: {
+    task: 'taxSlipT4eTask',
+    icon: '/images/dashboard/oas-tax-slip-icon.svg',
+    taskLink: 'taxSlipTaskLink',
+  },
+  ViewPaymentInfoIconWhite: {
+    task: 'viewPaymentInfo',
+    icon: '/images/dashboard/apply-for-benefit-icon-white.svg',
+    taskLink: 'recordOfEmployment',
+  },
+  //Seb tasks
+  ViewAgreementStatusDetails: {
+    task: 'viewAgreementStatus',
+    icon: '/images/dashboard/ei-view-documents-icon.svg',
+    taskLink: 'viewDocumentsLink',
+  },
+}
